+++ bit-set.d.ts
@@ -25,5 +25,5 @@ export default class BitSet implements Iterable<number> {
   entries(): Iterator<[number, number]>;
   [Symbol.iterator](): Iterator<number>;
   inspect(): any;
-  toJSON(): Uint32Array;
+  toJSON(): Array<number>;
 }